+++ lib/mino/tasks/builtin.clj
@@ -46,6 +46,7 @@
    "src/collections/vec.c" "src/collections/map.c"
    "src/collections/chunk.c"
    "src/collections/queue.c"
+   "src/collections/bytes.c"
    "src/collections/rbtree.c"
    "src/collections/builders.c"
    "src/collections/gc_handlers.c"