+++ src/eval/bc/vm.c
@@ -424,6 +424,7 @@ static mino_val *bc_protocol_type_disc(mino_state *S, mino_val *v)
     case MINO_AGENT:      return mino_keyword(S, "agent");
     case MINO_CHAN:       return mino_keyword(S, "chan");
     case MINO_QUEUE:      return mino_keyword(S, "queue");
+    case MINO_BYTES:      return mino_keyword(S, "bytes");
     }
     return mino_keyword(S, "unknown");
 }