+++ src/eval/special.c
@@ -1067,6 +1067,7 @@ mino_val *eval_impl(mino_state *S, mino_val *form, mino_env *env, int tail)
     case MINO_AGENT:
     case MINO_CHAN:
     case MINO_QUEUE:
+    case MINO_BYTES:
         return form;
     case MINO_SYMBOL:
         return eval_symbol(S, form, env);