+++ tests/run.clj
@@ -24,6 +24,7 @@
 (require "tests/jvm_statics_test")
 (require "tests/clojure_version_test")
 (require "tests/inst_test")
+(require "tests/bytes_test")
 (require "tests/macro_test")
 (require "tests/error_test")
 (require "tests/atom_test")